New netTensorRT version, cleaner, ready for debug, loads of questions.
